% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZRA.R
\name{plot.ZRA}
\alias{plot.ZRA}
\title{Ploting a ZRA-Class Obejct}
\usage{
\method{plot}{ZRA}(x,zero, ...)
}
\arguments{
\item{x}{(ZRA) An ZRA-Object.}

\item{zero}{(boolean) If zero=TRUE, they will be the 0 included in the Graph.}

\item{...}{further arguments passed to the plot method.}
}
\description{
This plot-Method will plot the original Time Series and the predicted Data, using dygraphs.
}

