% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZRA.R
\name{print.ZRA}
\alias{print.ZRA}
\title{Printig a ZRA-Class Obejct}
\usage{
\method{print}{ZRA}(x, ...)
}
\arguments{
\item{x}{(ZRA) An ZRA-Object.}

\item{...}{further arguments passed to the print method.}
}
\description{
This print-Method will print out the original Time Series and the predicted Data.
}

