% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ZRA.R
\name{ZRA}
\alias{ZRA}
\title{Dynamic Plots for Time Series Forecasting}
\usage{
ZRA(data, FP = 10, SL = c(0.8, 0.95), ...)
}
\arguments{
\item{data}{(ts) Time series data.}

\item{FP}{(numeric) Forecast period.}

\item{SL}{(numeric) significance levels.}

\item{...}{further arguments passed to the forecast-function.}
}
\value{
An Object of class "ZRA".
}
\description{
Dynamic Plots for Time Series Forecasting
}
\examples{
zra <- ZRA(fdeaths)
plot(zra, zero = TRUE)
}
\seealso{
forecast, dygraphs
}

